refactor: Completely refactor API and organise stuff to begin reworking property code.
BREAKING CHANGE: completely rehash API. prop namespace is now prop* prefixed functions. props() is
now split into getProps / setProps. All mixins are prefixed using with* (withRender, etc).
