--- src/component.js
@@ -1,3 +1,0 @@
-import { Component } from './mixins';
-
-export default Component();
