--- src/lifecycle/props-init.js
@@ -1,94 +1,0 @@
-import { _updateDebounced } from '../util/symbols';
-import data from '../util/data';
-import empty from '../util/empty';
-import getAttrMgr from '../util/attributes-manager';
-
-function getDefaultValue (elem, propDef) {
-  return typeof propDef.default === 'function'
-    ? propDef.default(elem, { name: propDef.nameOrSymbol })
-    : propDef.default;
-}
-
-function getPropData (elem, name) {
-  const elemData = data(elem, 'props');
-  return elemData[name] || (elemData[name] = {});
-}
-
-export function createNativePropertyDescriptor (propDef) {
-  const { nameOrSymbol } = propDef;
-
-  const prop = {
-    configurable: true,
-    enumerable: true
-  };
-
-  prop.beforeDefineProperty = elem => {
-    const propData = getPropData(elem, nameOrSymbol);
-    const attrSource = propDef.attrSource;
-
-    // Store attrSource name to property link.
-    if (attrSource) {
-      data(elem, 'attrSourceLinks')[attrSource] = nameOrSymbol;
-    }
-
-    // prop value before upgrading
-    let initialValue = elem[nameOrSymbol];
-
-    // Set up initial value if it wasn't specified.
-    let valueFromAttrSource = false;
-    if (empty(initialValue)) {
-      if (attrSource && elem.hasAttribute(attrSource)) {
-        valueFromAttrSource = true;
-        initialValue = propDef.deserialize(elem.getAttribute(attrSource));
-      } else {
-        initialValue = getDefaultValue(elem, propDef);
-      }
-    }
-
-    initialValue = propDef.coerce(initialValue);
-
-    propData.internalValue = initialValue;
-
-    // Reflect to Target Attribute
-    const mustReflect = propDef.attrTarget && !empty(initialValue) &&
-      (!valueFromAttrSource || propDef.attrTargetIsNotSource);
-
-    if (mustReflect) {
-      let serializedValue = propDef.serialize(initialValue);
-      getAttrMgr(elem).setAttrValue(propDef.attrTarget, serializedValue);
-    }
-  };
-
-  prop.get = function get () {
-    return getPropData(this, nameOrSymbol).internalValue;
-  };
-
-  prop.set = function set (newValue) {
-    const propData = getPropData(this, nameOrSymbol);
-    const useDefaultValue = empty(newValue);
-
-    if (useDefaultValue) {
-      newValue = getDefaultValue(this, propDef);
-    }
-
-    newValue = propDef.coerce(newValue);
-
-    // Queue a re-render.
-    this[_updateDebounced]();
-
-    // Update prop data so we can use it next time.
-    propData.internalValue = newValue;
-
-    // Reflect to Target attribute.
-    const mustReflect = propDef.attrTarget &&
-      (propDef.attrTargetIsNotSource || !propData.settingPropFromAttrSource);
-    if (mustReflect) {
-      // Note: setting the prop to empty implies the default value
-      // and therefore no attribute should be present!
-      let serializedValue = useDefaultValue ? null : propDef.serialize(newValue);
-      getAttrMgr(this).setAttrValue(propDef.attrTarget, serializedValue);
-    }
-  };
-
-  return prop;
-}
