+++ test/unit/index.spec.js
@@ -24,25 +24,24 @@ describe('exports', () => {
     expect(api.link).toBeA('function');
   });
 
-  it('skate.Mixins', () => {
-    expect(api.Mixins).toBeAn('object');
-    expect(api.Mixins.Component).toBeA('function', 'Component');
-    expect(api.Mixins.Props).toBeA('function', 'Props');
-    expect(api.Mixins.Raw).toBeA('function', 'Raw');
-    expect(api.Mixins.Render).toBeA('function', 'Render');
+  it('skate.with*', () => {
+    expect(api.withComponent).toBeA('function', 'withComponent');
+    expect(api.withProps).toBeA('function', 'withProps');
+    expect(api.withRaw).toBeA('function', 'withRaw');
+    expect(api.withRender).toBeA('function', 'withRender');
   });
 
-  it('skate.prop', () => {
-    expect(api.prop).toBeAn('object');
-    expect(api.prop.array).toBeAn('object', 'array');
-    expect(api.prop.boolean).toBeAn('object', 'boolean');
-    expect(api.prop.number).toBeAn('object', 'number');
-    expect(api.prop.object).toBeAn('object', 'object');
-    expect(api.prop.string).toBeAn('object', 'string');
+  it('skate.prop*', () => {
+    expect(api.propArray).toBeAn('object', 'array');
+    expect(api.propBoolean).toBeAn('object', 'boolean');
+    expect(api.propNumber).toBeAn('object', 'number');
+    expect(api.propObject).toBeAn('object', 'object');
+    expect(api.propString).toBeAn('object', 'string');
   });
 
-  it('skate.props', () => {
-    expect(api.props).toBeA('function');
+  it('skate.*Props', () => {
+    expect(api.getProps).toBeA('function');
+    expect(api.setProps).toBeA('function');
   });
 
   it('skate.h', () => {
