+++ test/unit/props.spec.js
@@ -5,7 +5,7 @@ import expect from 'expect';
 import afterMutations from '../lib/after-mutations';
 import fixture from '../lib/fixture';
 
-import { define, Mixins, props } from 'src';
+import { define, getProps, setProps, withProps } from 'src';
 import root from 'src/util/root';
 
 const { HTMLElement } = root;
@@ -14,7 +14,7 @@ describe('api/props', () => {
   let elem;
 
   beforeEach(done => {
-    elem = new (define(class extends Mixins.Props() {
+    elem = new (define(class extends withProps() {
       static get props () {
         return {
           prop1: null,
@@ -38,7 +38,7 @@ describe('api/props', () => {
 
   describe('getting', () => {
     it('should return only properties defined as props', () => {
-      const curr = props(elem);
+      const curr = getProps(elem);
 
       expect(curr.prop1).toEqual('test1');
       expect(curr.prop2).toEqual('test2');
@@ -49,7 +49,7 @@ describe('api/props', () => {
 
   describe('setting', () => {
     it('should set all properties', () => {
-      props(elem, {
+      setProps(elem, {
         prop1: 'updated1',
         prop2: 'updated2',
         undeclaredProp: 'updated3'
@@ -61,7 +61,7 @@ describe('api/props', () => {
 
     it('should asynchronously render if declared properties are set', done => {
       expect(elem._rendered).toEqual(1);
-      props(elem, { prop1: 'updated1' });
+      setProps(elem, { prop1: 'updated1' });
       afterMutations(
         () => expect(elem._rendered).toEqual(2),
         done
@@ -70,7 +70,7 @@ describe('api/props', () => {
 
     it('should asynchronously render once when multiple props are set', done => {
       expect(elem._rendered).toEqual(1);
-      props(elem, {
+      setProps(elem, {
         prop1: 'updated1',
         prop2: 'updated2'
       });
@@ -82,13 +82,13 @@ describe('api/props', () => {
 
     it('should not render if undeclared properties are set', () => {
       expect(elem._rendered).toEqual(1);
-      props(elem, { undeclaredProp: 'updated3' });
+      setProps(elem, { undeclaredProp: 'updated3' });
       expect(elem._rendered).toEqual(1);
     });
 
     it('should succeed on an uninitialised element', () => {
       const elem = new (define(class extends HTMLElement {}))();
-      props(elem, { undeclaredProp: 'foo' });
+      setProps(elem, { undeclaredProp: 'foo' });
       expect(elem.undeclaredProp).toBe('foo');
     });
   });
